+++ libs/cmcd/src/CMCD_MIME_TYPE.ts
@@ -1,0 +1,6 @@
+/**
+ * CMCD MIME type for event report payloads.
+ *
+ * @public
+ */
+export const CMCD_MIME_TYPE = 'application/cmcd'