+++ libs/cmcd/src/toCmcdHeaders.ts
@@ -12,6 +12,8 @@ import { toPreparedCmcdHeaders } from './toPreparedCmcdHeaders.ts'
  *
  * @returns The CMCD header shards.
  *
+ * @see {@link https://cta-wave.github.io/Resources/common-media-client-data--cta-5004-b.html#header-field-definition | CTA-5004-A Header Field Definition}
+ *
  * @public
  *
  * @example